Fixes for GCC4
git-svn-id: http://source.riscpkg.org/svn/RTK/trunk@429 fa0f773a-f822-0410-a9df-e7ef4170f6f5

+++ !RTK/rtk/desktop/basic_row_layout.cc
@@ -114,7 +114,7 @@ void basic_row_layout::remove_notify(component& c)
 
 basic_row_layout& basic_row_layout::cells(size_type xcells)
 {
-	for (size_type i=min(xcells,_components.size());i!=_components.size();++i)
+	for (size_type i=min((size_t)xcells,_components.size());i!=_components.size();++i)
 		if (component* c=_components[i]) c->remove();
 	_components.resize(xcells,0);
 	_xmin.resize(xcells+1,0);
